Migrate app stuff in util.ts to avoid circular dependencies.
I checked these methods were not used in webhooks or galaxy visualizations.
